Reorganize constants stracture
