+++ src/constants/routs.ts
@@ -1,0 +1,8 @@
+const AppRoute = {
+	root: '/',
+	login: '/login',
+	offer: '/offer/:id',
+	favorites: '/favorites',
+} as const;
+
+export {AppRoute};
